refactor: removed ElasticsearchIndexConstants.java Constants moved to the actual class where are most relevant
